refactor: cel libs mocks
